Move sample data into separate file
